Use abbreviated match
